requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
django-environ==0.11.2
whitenoise==6.7.0
djangorestframework==3.14.0
gunicorn==21.2.0
